fix: tests and felt value limit
