+++ crates/pathfinder/Cargo.toml
@@ -15,6 +15,7 @@ bigdecimal = { version = "0.3.0", features = ["serde"] }
 # paritys scale codec locks us here
 bitvec = "0.20.4"
 clap = "2.33.3"
+derive_more = "0.99.17"
 enum-iterator = "0.7.0"
 hex = "0.4.3"
 home = "0.5.3"
