+++ crates/pathfinder/src/core.rs
@@ -2,12 +2,13 @@
 //! home of their own.
 //!
 //! This includes many trivial wrappers around [StarkHash] which help by providing additional type safety.
+use derive_more::{From, FromStr};
 use pedersen::StarkHash;
 use serde::{Deserialize, Serialize};
 use web3::types::H256;
 
 /// The address of a StarkNet contract.
-#[derive(Debug, Copy, Clone, PartialEq, Deserialize, Serialize)]
+#[derive(Copy, Clone, Debug, Deserialize, From, FromStr, PartialEq, Serialize)]
 pub struct ContractAddress(pub StarkHash);
 
 /// A StarkNet contract's hash. This is a hash over a contract's
@@ -31,11 +32,11 @@ pub struct ContractStateHash(pub StarkHash);
 pub struct ContractRoot(pub StarkHash);
 
 /// Entry point of a StarkNet `call`.
-#[derive(Debug, Copy, Clone, PartialEq, Deserialize, Serialize)]
+#[derive(Copy, Clone, Debug, Deserialize, From, FromStr, PartialEq, Serialize)]
 pub struct EntryPoint(pub StarkHash);
 
 /// A single parameter passed to a StarkNet `call`.
-#[derive(Debug, Copy, Clone, PartialEq, Deserialize, Serialize)]
+#[derive(Copy, Clone, Debug, Deserialize, From, FromStr, PartialEq, Serialize)]
 pub struct CallParam(pub StarkHash);
 
 /// A single result value of a StarkNet `call`.
@@ -60,7 +61,7 @@ pub struct StorageValue(pub StarkHash);
 pub struct GlobalRoot(pub StarkHash);
 
 /// A StarkNet block hash.
-#[derive(Debug, Default, Copy, Clone, PartialEq, Deserialize, Serialize)]
+#[derive(Copy, Clone, Debug, Default, From, FromStr, PartialEq, Deserialize, Serialize)]
 pub struct StarknetBlockHash(pub StarkHash);
 
 /// A StarkNet block number.
@@ -72,7 +73,7 @@ pub struct StarknetBlockNumber(pub u64);
 pub struct StarknetBlockTimestamp(pub u64);
 
 /// A StarkNet transaction hash.
-#[derive(Debug, Copy, Clone, PartialEq, Deserialize, Serialize)]
+#[derive(Copy, Clone, Debug, Deserialize, From, FromStr, PartialEq, Serialize)]
 pub struct StarknetTransactionHash(pub StarkHash);
 
 /// A StarkNet transaction hash.
