+++ crates/pathfinder/src/rpc.rs
@@ -258,7 +258,7 @@ mod tests {
 
     lazy_static::lazy_static! {
         static ref GENESIS_BLOCK_HASH: H256 = H256::from_str("0x07d328a71faf48c5c3857e99f20a77b18522480956d1cd5bff1ff2df3c8b427b").unwrap();
-        static ref INVALID_BLOCK_HASH: H256 = H256::from_str("0x13d8d8bb5716cd3f16e54e3a6ff1a50542461d9022e5f4dec7a4b064041ab8d7").unwrap();
+        static ref INVALID_BLOCK_HASH: H256 = H256::from_str("0x06d328a71faf48c5c3857e99f20a77b18522480956d1cd5bff1ff2df3c8b427b").unwrap();
         static ref UNKNOWN_BLOCK_HASH: H256 = H256::from_str("0x03c85a69453e63fd475424ecc70438bd855cd76e6f0d5dec0d0dd56e0f7a771c").unwrap();
         static ref PRE_DEPLOY_CONTRACT_BLOCK_HASH: H256 = H256::from_str("0x05ef884a311df4339c8df791ce19bf305d7cf299416666b167bc56dd2d1f435f").unwrap();
         static ref DEPLOY_CONTRACT_BLOCK_HASH: H256 = H256::from_str("0x07177acba67cb659e336abb3a158c8d29770b87b1b62e2bfa94cd376b72d34c5").unwrap();
@@ -271,6 +271,28 @@ mod tests {
         static ref UNKNOWN_CONTRACT_ADDR: relaxed::H256 = H256::from_str("0x0739636829ad5205d81af792a922a40e35c0ec7a72f4859843ee2e2a0d6f0af0").unwrap().into();
         static ref VALID_ENTRY_POINT: relaxed::H256 = H256::from_str("0x0362398bec32bc0ebb411203221a35a0301193a96f317ebe5e40be9f60d15320").unwrap().into();
         static ref LOCALHOST: SocketAddr = SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 0));
+    }
+
+    mod v2 {
+        use crate::core::{
+            ContractAddress, EntryPoint, StarknetBlockHash, StarknetTransactionHash,
+        };
+        use std::str::FromStr;
+
+        lazy_static::lazy_static! {
+            pub static ref GENESIS_BLOCK_HASH: StarknetBlockHash = StarknetBlockHash::from_str("0x07d328a71faf48c5c3857e99f20a77b18522480956d1cd5bff1ff2df3c8b427b").unwrap();
+            pub static ref INVALID_BLOCK_HASH: StarknetBlockHash = StarknetBlockHash::from_str("0x06d328a71faf48c5c3857e99f20a77b18522480956d1cd5bff1ff2df3c8b427b").unwrap();
+            pub static ref UNKNOWN_BLOCK_HASH: StarknetBlockHash = StarknetBlockHash::from_str("0x03c85a69453e63fd475424ecc70438bd855cd76e6f0d5dec0d0dd56e0f7a771c").unwrap();
+            pub static ref CONTRACT_BLOCK_HASH: StarknetBlockHash = StarknetBlockHash::from_str("0x03871c8a0c3555687515a07f365f6f5b1d8c2ae953f7844575b8bde2b2efed27").unwrap();
+            pub static ref VALID_TX_HASH: StarknetTransactionHash = StarknetTransactionHash::from_str("0x0493d8fab73af67e972788e603aee18130facd3c7685f16084ecd98b07153e24").unwrap();
+            pub static ref INVALID_TX_HASH: StarknetTransactionHash = StarknetTransactionHash::from_str("0x0393d8fab73af67e972788e603aee18130facd3c7685f16084ecd98b07153e24").unwrap();
+            pub static ref UNKNOWN_TX_HASH: StarknetTransactionHash = StarknetTransactionHash::from_str("0x015e4bb72df94be3044139fea2116c4d54add05cf9ef8f35aea114b5cea94713").unwrap();
+            pub static ref VALID_CONTRACT_ADDR: ContractAddress = ContractAddress::from_str("0x06fbd460228d843b7fbef670ff15607bf72e19fa94de21e29811ada167b4ca39").unwrap();
+            pub static ref INVALID_CONTRACT_ADDR: ContractAddress = ContractAddress::from_str("0x05fbd460228d843b7fbef670ff15607bf72e19fa94de21e29811ada167b4ca39").unwrap();
+            pub static ref UNKNOWN_CONTRACT_ADDR: ContractAddress = ContractAddress::from_str("0x0739636829ad5205d81af792a922a40e35c0ec7a72f4859843ee2e2a0d6f0af0").unwrap();
+            pub static ref VALID_ENTRY_POINT: EntryPoint = EntryPoint::from_str("0x0362398bec32bc0ebb411203221a35a0301193a96f317ebe5e40be9f60d15320").unwrap();
+            pub static ref INVALID_ENTRY_POINT: EntryPoint = EntryPoint::from_str("").unwrap();
+        }
     }
 
     mod error {
@@ -1070,11 +1092,13 @@ mod tests {
 
     mod call {
         use super::*;
-        use crate::rpc::types::{request::Call, BlockHashOrTag, Tag};
-        use web3::types::H256;
+        use crate::{
+            core::CallParam,
+            rpc::types::{request::Call, BlockHashOrTag, Tag},
+        };
 
         lazy_static::lazy_static! {
-            static ref CALL_DATA: Vec<H256> = vec![H256::from_str("0x0000000000000000000000000000000000000000000000000000000000001234").unwrap()];
+            static ref CALL_DATA: Vec<CallParam> = vec![CallParam::from_str("1234").unwrap()];
         }
 
         #[tokio::test]
@@ -1082,8 +1106,8 @@ mod tests {
             let params = rpc_params!(
                 Call {
                     calldata: CALL_DATA.clone(),
-                    contract_address: **VALID_CONTRACT_ADDR,
-                    entry_point_selector: **VALID_ENTRY_POINT,
+                    contract_address: *v2::VALID_CONTRACT_ADDR,
+                    entry_point_selector: *v2::VALID_ENTRY_POINT,
                 },
                 BlockHashOrTag::Hash(*INVOKE_CONTRACT_BLOCK_HASH)
             );
@@ -1100,8 +1124,8 @@ mod tests {
                 let params = rpc_params!(
                     Call {
                         calldata: CALL_DATA.clone(),
-                        contract_address: **VALID_CONTRACT_ADDR,
-                        entry_point_selector: **VALID_ENTRY_POINT,
+                        contract_address: *v2::VALID_CONTRACT_ADDR,
+                        entry_point_selector: *v2::VALID_ENTRY_POINT,
                     },
                     BlockHashOrTag::Tag(Tag::Latest)
                 );
@@ -1134,8 +1158,8 @@ mod tests {
             let params = rpc_params!(
                 Call {
                     calldata: CALL_DATA.clone(),
-                    contract_address: **VALID_CONTRACT_ADDR,
-                    entry_point_selector: **VALID_ENTRY_POINT,
+                    contract_address: *v2::VALID_CONTRACT_ADDR,
+                    entry_point_selector: *v2::VALID_ENTRY_POINT,
                 },
                 BlockHashOrTag::Tag(Tag::Pending)
             );
@@ -1149,8 +1173,8 @@ mod tests {
             let params = rpc_params!(
                 Call {
                     calldata: CALL_DATA.clone(),
-                    contract_address: **VALID_CONTRACT_ADDR,
-                    entry_point_selector: H256::zero(),
+                    contract_address: *v2::VALID_CONTRACT_ADDR,
+                    entry_point_selector: *v2::INVALID_ENTRY_POINT,
                 },
                 BlockHashOrTag::Tag(Tag::Latest)
             );
@@ -1168,8 +1192,8 @@ mod tests {
             let params = rpc_params!(
                 Call {
                     calldata: CALL_DATA.clone(),
-                    contract_address: **INVALID_CONTRACT_ADDR,
-                    entry_point_selector: **VALID_ENTRY_POINT,
+                    contract_address: *v2::INVALID_CONTRACT_ADDR,
+                    entry_point_selector: *v2::VALID_ENTRY_POINT,
                 },
                 BlockHashOrTag::Tag(Tag::Latest)
             );
@@ -1187,8 +1211,8 @@ mod tests {
             let params = rpc_params!(
                 Call {
                     calldata: vec![],
-                    contract_address: **VALID_CONTRACT_ADDR,
-                    entry_point_selector: **VALID_ENTRY_POINT,
+                    contract_address: *v2::VALID_CONTRACT_ADDR,
+                    entry_point_selector: *v2::VALID_ENTRY_POINT,
                 },
                 BlockHashOrTag::Tag(Tag::Latest)
             );
@@ -1206,8 +1230,8 @@ mod tests {
             let params = rpc_params!(
                 Call {
                     calldata: CALL_DATA.clone(),
-                    contract_address: **VALID_CONTRACT_ADDR,
-                    entry_point_selector: **VALID_ENTRY_POINT,
+                    contract_address: *v2::VALID_CONTRACT_ADDR,
+                    entry_point_selector: *v2::VALID_ENTRY_POINT,
                 },
                 BlockHashOrTag::Hash(*PRE_DEPLOY_CONTRACT_BLOCK_HASH)
             );
@@ -1225,8 +1249,8 @@ mod tests {
             let params = rpc_params!(
                 Call {
                     calldata: CALL_DATA.clone(),
-                    contract_address: **VALID_CONTRACT_ADDR,
-                    entry_point_selector: **VALID_ENTRY_POINT,
+                    contract_address: *v2::VALID_CONTRACT_ADDR,
+                    entry_point_selector: *v2::VALID_ENTRY_POINT,
                 },
                 BlockHashOrTag::Hash(*INVALID_BLOCK_HASH)
             );
@@ -1244,8 +1268,8 @@ mod tests {
             let params = rpc_params!(
                 Call {
                     calldata: CALL_DATA.clone(),
-                    contract_address: **VALID_CONTRACT_ADDR,
-                    entry_point_selector: **VALID_ENTRY_POINT,
+                    contract_address: *v2::VALID_CONTRACT_ADDR,
+                    entry_point_selector: *v2::VALID_ENTRY_POINT,
                 },
                 BlockHashOrTag::Hash(*UNKNOWN_BLOCK_HASH)
             );
