+++ crates/pathfinder/src/sequencer.rs
@@ -55,6 +55,7 @@ where
     // Status codes <400;499> and <501;599> are mapped to SequencerError::TransportError
     resp.error_for_status_ref().map(|_| ())?;
     let resp = resp.text().await?;
+    eprintln!("{}", resp);
     // Attempt to deserialize the actual data we are looking for
     let deserialized = serde_json::from_str::<T>(resp.as_str())?;
     Ok(deserialized)
@@ -229,7 +230,7 @@ mod tests {
 
     lazy_static::lazy_static! {
         static ref GENESIS_BLOCK_HASH: H256 = H256::from_str("0x07d328a71faf48c5c3857e99f20a77b18522480956d1cd5bff1ff2df3c8b427b").unwrap();
-        static ref INVALID_BLOCK_HASH: H256 = H256::from_str("0x13d8d8bb5716cd3f16e54e3a6ff1a50542461d9022e5f4dec7a4b064041ab8d7").unwrap();
+        static ref INVALID_BLOCK_HASH: H256 = H256::from_str("0x06d328a71faf48c5c3857e99f20a77b18522480956d1cd5bff1ff2df3c8b427b").unwrap();
         static ref UNKNOWN_BLOCK_HASH: H256 = H256::from_str("0x03c85a69453e63fd475424ecc70438bd855cd76e6f0d5dec0d0dd56e0f7a771c").unwrap();
         static ref CONTRACT_BLOCK_HASH: H256 = H256::from_str("0x03871c8a0c3555687515a07f365f6f5b1d8c2ae953f7844575b8bde2b2efed27").unwrap();
         static ref VALID_TX_HASH: H256 = H256::from_str("0x0493d8fab73af67e972788e603aee18130facd3c7685f16084ecd98b07153e24").unwrap();
@@ -239,6 +240,28 @@ mod tests {
         static ref INVALID_CONTRACT_ADDR: H256 = H256::from_str("0x16fbd460228d843b7fbef670ff15607bf72e19fa94de21e29811ada167b4ca39").unwrap();
         static ref UNKNOWN_CONTRACT_ADDR: H256 = H256::from_str("0x0739636829ad5205d81af792a922a40e35c0ec7a72f4859843ee2e2a0d6f0af0").unwrap();
         static ref VALID_ENTRY_POINT: H256 = H256::from_str("0x0362398bec32bc0ebb411203221a35a0301193a96f317ebe5e40be9f60d15320").unwrap();
+    }
+
+    mod v2 {
+        use crate::core::{
+            ContractAddress, EntryPoint, StarknetBlockHash, StarknetTransactionHash,
+        };
+        use std::str::FromStr;
+
+        lazy_static::lazy_static! {
+            pub static ref GENESIS_BLOCK_HASH: StarknetBlockHash = StarknetBlockHash::from_str("0x07d328a71faf48c5c3857e99f20a77b18522480956d1cd5bff1ff2df3c8b427b").unwrap();
+            pub static ref INVALID_BLOCK_HASH: StarknetBlockHash = StarknetBlockHash::from_str("0x06d328a71faf48c5c3857e99f20a77b18522480956d1cd5bff1ff2df3c8b427b").unwrap();
+            pub static ref UNKNOWN_BLOCK_HASH: StarknetBlockHash = StarknetBlockHash::from_str("0x03c85a69453e63fd475424ecc70438bd855cd76e6f0d5dec0d0dd56e0f7a771c").unwrap();
+            pub static ref CONTRACT_BLOCK_HASH: StarknetBlockHash = StarknetBlockHash::from_str("0x03871c8a0c3555687515a07f365f6f5b1d8c2ae953f7844575b8bde2b2efed27").unwrap();
+            pub static ref VALID_TX_HASH: StarknetTransactionHash = StarknetTransactionHash::from_str("0x0493d8fab73af67e972788e603aee18130facd3c7685f16084ecd98b07153e24").unwrap();
+            pub static ref INVALID_TX_HASH: StarknetTransactionHash = StarknetTransactionHash::from_str("0x0393d8fab73af67e972788e603aee18130facd3c7685f16084ecd98b07153e24").unwrap();
+            pub static ref UNKNOWN_TX_HASH: StarknetTransactionHash = StarknetTransactionHash::from_str("0x015e4bb72df94be3044139fea2116c4d54add05cf9ef8f35aea114b5cea94713").unwrap();
+            pub static ref VALID_CONTRACT_ADDR: ContractAddress = ContractAddress::from_str("0x06fbd460228d843b7fbef670ff15607bf72e19fa94de21e29811ada167b4ca39").unwrap();
+            pub static ref INVALID_CONTRACT_ADDR: ContractAddress = ContractAddress::from_str("0x05fbd460228d843b7fbef670ff15607bf72e19fa94de21e29811ada167b4ca39").unwrap();
+            pub static ref UNKNOWN_CONTRACT_ADDR: ContractAddress = ContractAddress::from_str("0x0739636829ad5205d81af792a922a40e35c0ec7a72f4859843ee2e2a0d6f0af0").unwrap();
+            pub static ref VALID_ENTRY_POINT: EntryPoint = EntryPoint::from_str("0x0362398bec32bc0ebb411203221a35a0301193a96f317ebe5e40be9f60d15320").unwrap();
+            pub static ref INVALID_ENTRY_POINT: EntryPoint = EntryPoint::from_str("").unwrap();
+        }
     }
 
     /// Convenience wrapper
@@ -434,7 +457,7 @@ mod tests {
             .unwrap_err();
             assert_matches!(
                 error,
-                SequencerError::StarknetError(e) => assert_eq!(e.code, StarknetErrorCode::OutOfRangeBlockHash)
+                SequencerError::StarknetError(e) => assert_eq!(e.code, StarknetErrorCode::BlockNotFound)
             );
         }
     }
@@ -499,8 +522,8 @@ mod tests {
                     .call(
                         request::Call {
                             calldata: vec![U256::from(1234u64)],
-                            contract_address: *VALID_CONTRACT_ADDR,
-                            entry_point_selector: H256::zero(),
+                            contract_address: *v2::VALID_CONTRACT_ADDR,
+                            entry_point_selector: *v2::INVALID_ENTRY_POINT,
                             signature: vec![],
                         },
                         BlockHashOrTag::Tag(Tag::Latest),
@@ -521,8 +544,8 @@ mod tests {
                     .call(
                         request::Call {
                             calldata: vec![U256::from(1234u64)],
-                            contract_address: *INVALID_CONTRACT_ADDR,
-                            entry_point_selector: *VALID_ENTRY_POINT,
+                            contract_address: *v2::INVALID_CONTRACT_ADDR,
+                            entry_point_selector: *v2::VALID_ENTRY_POINT,
                             signature: vec![],
                         },
                         BlockHashOrTag::Tag(Tag::Latest),
@@ -532,7 +555,7 @@ mod tests {
             .unwrap_err();
             assert_matches!(
                 error,
-                SequencerError::StarknetError(e) => assert_eq!(e.code, StarknetErrorCode::OutOfRangeContractAddress)
+                SequencerError::StarknetError(e) => assert_eq!(e.code, StarknetErrorCode::UninitializedContract)
             );
         }
 
@@ -543,8 +566,8 @@ mod tests {
                     .call(
                         request::Call {
                             calldata: vec![],
-                            contract_address: *VALID_CONTRACT_ADDR,
-                            entry_point_selector: *VALID_ENTRY_POINT,
+                            contract_address: *v2::VALID_CONTRACT_ADDR,
+                            entry_point_selector: *v2::VALID_ENTRY_POINT,
                             signature: vec![],
                         },
                         BlockHashOrTag::Hash(*CONTRACT_BLOCK_HASH),
@@ -565,8 +588,8 @@ mod tests {
                     .call(
                         request::Call {
                             calldata: vec![],
-                            contract_address: *VALID_CONTRACT_ADDR,
-                            entry_point_selector: *VALID_ENTRY_POINT,
+                            contract_address: *v2::VALID_CONTRACT_ADDR,
+                            entry_point_selector: *v2::VALID_ENTRY_POINT,
                             signature: vec![],
                         },
                         BlockHashOrTag::Hash(*GENESIS_BLOCK_HASH),
@@ -587,8 +610,8 @@ mod tests {
                     .call(
                         request::Call {
                             calldata: vec![U256::from(1234u64)],
-                            contract_address: *VALID_CONTRACT_ADDR,
-                            entry_point_selector: *VALID_ENTRY_POINT,
+                            contract_address: *v2::VALID_CONTRACT_ADDR,
+                            entry_point_selector: *v2::VALID_ENTRY_POINT,
                             signature: vec![],
                         },
                         BlockHashOrTag::Hash(*INVALID_BLOCK_HASH),
@@ -598,7 +621,7 @@ mod tests {
             .unwrap_err();
             assert_matches!(
                 error,
-                SequencerError::StarknetError(e) => assert_eq!(e.code, StarknetErrorCode::OutOfRangeBlockHash)
+                SequencerError::StarknetError(e) => assert_eq!(e.code, StarknetErrorCode::BlockNotFound)
             );
         }
 
@@ -609,8 +632,8 @@ mod tests {
                     .call(
                         request::Call {
                             calldata: vec![U256::from(1234u64)],
-                            contract_address: *VALID_CONTRACT_ADDR,
-                            entry_point_selector: *VALID_ENTRY_POINT,
+                            contract_address: *v2::VALID_CONTRACT_ADDR,
+                            entry_point_selector: *v2::VALID_ENTRY_POINT,
                             signature: vec![],
                         },
                         BlockHashOrTag::Hash(*UNKNOWN_BLOCK_HASH),
@@ -631,8 +654,8 @@ mod tests {
                     .call(
                         request::Call {
                             calldata: vec![U256::from(1234u64)],
-                            contract_address: *VALID_CONTRACT_ADDR,
-                            entry_point_selector: *VALID_ENTRY_POINT,
+                            contract_address: *v2::VALID_CONTRACT_ADDR,
+                            entry_point_selector: *v2::VALID_ENTRY_POINT,
                             signature: vec![],
                         },
                         BlockHashOrTag::Hash(*CONTRACT_BLOCK_HASH),
@@ -649,8 +672,8 @@ mod tests {
                     .call(
                         request::Call {
                             calldata: vec![U256::from(1234u64)],
-                            contract_address: *VALID_CONTRACT_ADDR,
-                            entry_point_selector: *VALID_ENTRY_POINT,
+                            contract_address: *v2::VALID_CONTRACT_ADDR,
+                            entry_point_selector: *v2::VALID_ENTRY_POINT,
                             signature: vec![],
                         },
                         BlockHashOrTag::Tag(Tag::Latest),
@@ -667,8 +690,8 @@ mod tests {
                     .call(
                         request::Call {
                             calldata: vec![U256::from(1234u64)],
-                            contract_address: *VALID_CONTRACT_ADDR,
-                            entry_point_selector: *VALID_ENTRY_POINT,
+                            contract_address: *v2::VALID_CONTRACT_ADDR,
+                            entry_point_selector: *v2::VALID_ENTRY_POINT,
                             signature: vec![],
                         },
                         BlockHashOrTag::Tag(Tag::Pending),
@@ -720,7 +743,7 @@ mod tests {
             .unwrap_err();
             assert_matches!(
                 error,
-                SequencerError::StarknetError(e) => assert_eq!(e.code, StarknetErrorCode::OutOfRangeBlockHash)
+                SequencerError::StarknetError(e) => assert_eq!(e.code, StarknetErrorCode::BlockNotFound)
             );
         }
 
@@ -851,7 +874,7 @@ mod tests {
             .unwrap_err();
             assert_matches!(
                 error,
-                SequencerError::StarknetError(e) => assert_eq!(e.code, StarknetErrorCode::OutOfRangeBlockHash)
+                SequencerError::StarknetError(e) => assert_eq!(e.code, StarknetErrorCode::BlockNotFound)
             );
         }
