+++ crates/pathfinder/src/storage/contract.rs
@@ -101,6 +101,7 @@ impl ContractsTable {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::str::FromStr;
 
     #[test]
     fn get_hash() {
